:truck: [#433] -- Re-organize date field stories/implementation
Split the stories and implementation by widget, tackling the datepicker
first and preparing for the input group variant.
