+++ src/components/forms/DateField/DateField.mdx
@@ -1,14 +1,14 @@
-import {ArgTypes, Canvas, Meta, Story} from '@storybook/blocks';
+import {Meta} from '@storybook/blocks';
 
-import * as DateFieldStories from './DateField.stories';
+<Meta title="Pure React Components / Forms / DateField" />
 
-<Meta of={DateFieldStories} />
+# Date Field
 
-## Date field
+The date field supports multiple widgets.
 
-We identify different types of date inputs, each coming with their most-appropriate UI element.
+## Types of dates
 
-**Nearby dates**
+### Nearby dates
 
 Think of appointment dates, booking dates and similar. Here, the expected date is in the nearby
 future or past. A calendar component is appropriate because:
@@ -17,7 +17,9 @@ future or past. A calendar component is appropriate because:
 - you can indicate the current date
 - it is not expected that the user needs to click the year/month button too many times
 
-**Known dates (currently not supported)**
+See the `Datepicker` stories for this.
+
+### Known dates
 
 Known dates are exact dates that the user typically knows of the top of their head, such as birth
 dates, credit card expiry dates...
@@ -25,39 +27,15 @@ dates, credit card expiry dates...
 Calendars are not suitable due to the usually bad default/initial value. Preferably they are able to
 enter the numeric value for day, month and year, leveraging autocomplete as well.
 
-On mobile devices, you'd typically want to use the native date widget too.
+On mobile devices, you'd typically want to use the native date widget too. TODO: is this statement
+correct?
 
-### Datepicker/calendar
-
-The datepicker/calendar is suitable for nearby dates. We aim to make it accessible to assistive
-technology and keyboard-only navigation.
-
-Known issues/enhancements:
-
-- Use arrow keys to navigate inside calendar
-- Selecting a date in the next/previous month has some odd behaviour at the moment
-
-<Canvas of={DateFieldStories.Datepicker} />
-
-<Canvas of={DateFieldStories.LimitedRangeDatepicker} />
-
-<Canvas of={DateFieldStories.DisabledDatesDatepicker} />
-
-## Props
-
-<ArgTypes of={DateFieldStories} />
+See the `Input group` stories for this.
 
 ## References
 
-- [GOV.UK](https://design-system.service.gov.uk/patterns/dates/)
-- [Blogpost](https://medium.com/samsung-internet-dev/making-input-type-date-complicated-a544fd27c45a)
-- [nl-design-system/backlog#189](https://github.com/nl-design-system/backlog/issues/189)
-- [nl-design-system/backlog#188](https://github.com/nl-design-system/backlog/issues/188)
-- [nl-design-system/backlog#35](https://github.com/nl-design-system/backlog/issues/35#issuecomment-1547704753)
-
-## No asterisks
-
-The backend can be configured to treat fields as required by default and instead mark optional
-fields explicitly.
-
-<Story of={DateFieldStories.NoAsterisks} />
+- https://design-system.service.gov.uk/patterns/dates/
+- https://medium.com/samsung-internet-dev/making-input-type-date-complicated-a544fd27c45a
+- https://github.com/nl-design-system/backlog/issues/189
+- https://github.com/nl-design-system/backlog/issues/188
+- https://github.com/nl-design-system/backlog/issues/35#issuecomment-1547704753
